backend/pyproject.toml
[project]
name = "betterssb-backend"
version = "0.1.0"
description = "BetterSSB API server — schedule optimization, RMP proxy, chatbot, and transcript analysis"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "asyncpg>=0.30.0",
    "alembic>=1.15.0",
    "openai>=1.60.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.7.0",
    "beautifulsoup4>=4.13.0",
    "lxml>=5.3.0",
    "pyjwt[crypto]>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0",
    "aiosqlite>=0.21.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
]
ignore = [
    "E501",  # line length handled by formatter
    "B008",  # function call in default argument (Depends)
    "TCH",   # type-checking import blocks — overkill for this project size
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
